aula8update

+++ backend/src/todos/use-cases/delete-todo.use-cases.ts
@@ -1,17 +1,25 @@
-import { Inject, Injectable, Logger } from "@nestjs/common";
-import { DeleteTodoRepository} from "../repository";
+import { Inject, Injectable, Logger, NotFoundException } from "@nestjs/common";
+import { DeleteTodoRepository, FindTodoByIdRepository} from "../repository";
 
 @Injectable ()
 export class DeleteTodoUseCases {
     constructor (
+        private readonly findTodoByIdRepository: FindTodoByIdRepository,
         private readonly deleteTodoRepository: DeleteTodoRepository,
         private readonly logger: Logger,
     ) {}
 
     async delete(id: string) {
         try {
-            this.logger.log('Fetching all todos...')
-            const todo = await this.deleteTodoRepository.delete(id);
+            this.logger.log('Delecting toDo...')
+
+            const todo = await this.findTodoByIdRepository.findById(id);
+            
+            if (!todo) {
+                throw new NotFoundException('ToDo not found')
+            }
+
+            await this.deleteTodoRepository.delete(id)
             this.logger.log('Todos fetched successfully!');
             return todo;
         } catch (error) {
@@ -19,4 +27,4 @@ export class DeleteTodoUseCases {
             throw new Error('Failed to fetched todos');
         }
     }
-}
+}
